Adapt to trajectory task
Adapt the example gym environment for the trajectory task and provide
separate environments for the simulation and the real robot phases
(sharing common code in a base class).

Further add an example policy for trajectory task.  The example policy
does not actually attempt to grasp the cube but instead simply points to
the goal positions with one finger tip.